sklearn/hpsearch/working_requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
google-auth==1.1.0
idna==2.6
ipaddress==1.0.18
kubernetes==3.0.0
pyasn1==0.3.5
pyasn1-modules==0.1.4
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
websocket-client==0.40.0
